Clean up, move pill button component to shared
