+++ src/shared/models/statusLabels.ts
@@ -4,4 +4,4 @@ export const statusLabels: Record<WorkStatus, string> = {
     looking: "Currently looking for work",
     passive: "Passively looking for work",
     not_looking: "Don't want to hear about work",
-  };
+};
